plugins/blur: Remove contrast effect leftovers
It looks like something went wrong with backport patches and there is
still the backgroundcontrast effect around. On top of it, the blur
effect applies its own hardcoded contrast effect too.

We could remove the backgroundcontrast effect and backport relevant blur
effect changes from kwin_wayland, but kwin-x11 with backgroundcontrast
is already shipped and it's too late to flip flop with that now, it's
better to leave things as is and remove the hardcoded part from the blur
effect.